rename vtk_domain_writer.h to vtk_writer.h
